+++ lib/src/apps/app_stores/app_stores.dart
@@ -3,4 +3,5 @@ export 'huawei_app_gallery_store.dart';
 export 'ios_app_store.dart';
 export 'mac_app_store.dart';
 export 'microsoft_store.dart';
+export 'myket_store.dart';
 export 'play_store.dart';