+++ lib/src/apps/app_stores/cafe_bazaar_store.dart
@@ -2,7 +2,7 @@ import 'dart:core';
 
 import 'package:deeplink_x/src/core/app_actions/store_app_action.dart';
 import 'package:deeplink_x/src/core/enums/action_type_enum.dart';
-import 'package:deeplink_x/src/core/enums/platform_enum.dart';
+import 'package:deeplink_x/src/core/enums/platform_type.dart';
 
 /// Cafe Bazaar-specific action types that define available deeplink actions
 enum CafeBazaarActionType implements ActionTypeEnum {
@@ -25,7 +25,7 @@ class CafeBazaarAction extends StoreAppAction {
   }) : super(actionType: type, platform: platformType);
 
   /// The native platform type
-  static const platformType = PlatformEnum.android;
+  static const platformType = PlatformType.android;
 
   /// Base URI for Cafe Bazaar app deeplinks
   static const baseUrl = 'bazaar://';