+++ lib/src/apps/app_stores/ios_app_store.dart
@@ -2,7 +2,7 @@ import 'dart:core';
 
 import 'package:deeplink_x/src/core/app_actions/store_app_action.dart';
 import 'package:deeplink_x/src/core/enums/action_type_enum.dart';
-import 'package:deeplink_x/src/core/enums/platform_enum.dart';
+import 'package:deeplink_x/src/core/enums/platform_type.dart';
 
 /// iOS App Store-specific action types that define available deeplink actions
 enum IOSAppStoreActionType implements ActionTypeEnum {
@@ -31,7 +31,7 @@ class IOSAppStoreAction extends StoreAppAction {
   }) : super(actionType: type, platform: platformType);
 
   /// The native platform type
-  static const platformType = PlatformEnum.ios;
+  static const platformType = PlatformType.ios;
 
   /// Base URI for App Store app deeplinks
   static const baseUrl = 'itms-apps://itunes.apple.com';