+++ lib/src/core/deeplink_x.dart
@@ -1,8 +1,7 @@
-import 'dart:io';
-
 import 'package:deeplink_x/src/core/app_actions/app_action.dart';
-import 'package:deeplink_x/src/core/enums/platform_enum.dart';
+import 'package:deeplink_x/src/core/enums/platform_type.dart';
 import 'package:deeplink_x/src/utils/launcher_util.dart';
+import 'package:deeplink_x/src/utils/platform_util.dart';
 
 /// A class to handle deeplink actions across different platforms
 class DeeplinkX {
@@ -13,13 +12,13 @@ class DeeplinkX {
   /// - [platformType]: Optional platform type override. If not provided, automatically detects from current operating system.
   DeeplinkX({
     final LauncherUtil? launcherUtil,
-    final PlatformEnum? platformType,
+    final PlatformType? platformType,
   })  : _launcherUtil = launcherUtil ?? const LauncherUtil(),
-        _playformType = platformType ?? PlatformEnum.fromOperatingSystem(Platform.operatingSystem);
+        _platformType = platformType ?? const PlatformUtil().getCurrentPlatform();
 
   final LauncherUtil _launcherUtil;
 
-  final PlatformEnum _playformType;
+  final PlatformType _platformType;
 
   /// Launches a deeplink action.
   ///
@@ -30,7 +29,7 @@ class DeeplinkX {
   /// * `true` if the action was successfully launched
   /// * `false` if the action could not be launched or no URIs were available
   Future<bool> launchAction(final AppAction action) async {
-    final uris = await action.getUris(_playformType);
+    final uris = await action.getUris(_platformType);
     if (uris.isEmpty) {
       return false;
     }
@@ -66,7 +65,7 @@ class DeeplinkX {
   /// );
   /// ```
   Future<bool> canLaunch(final AppAction action) async {
-    final uris = await action.getUris(_playformType);
+    final uris = await action.getUris(_platformType);
     if (uris.isEmpty) {
       return false;
     }
@@ -78,5 +77,31 @@ class DeeplinkX {
       }
     }
     return false;
+  }
+
+  /// Checks app is installed and if a native deeplink URI can be launched.
+  ///
+  /// Takes an [action] parameter of type [AppAction] and checks if its
+  /// native URI can be launched.
+  ///
+  /// Returns a [Future<bool>] that completes with:
+  /// * `true` if the native URI can be launched
+  /// * `false` if the native URI cannot be launched
+  ///
+  /// Note: apps schemes must be registered based on the platform.
+  /// Check Platform-Specific Configurations in apps documents for more details.
+  ///
+  /// Example:
+  /// ```dart
+  /// final deeplinkX = DeeplinkX();
+  ///
+  /// // Check if native Instagram URI can be launched
+  /// final canLaunchNative = await deeplinkX.canLaunchNativeDeeplink(
+  ///   Instagram.openProfile('username')
+  /// );
+  /// ```
+  Future<bool> canLaunchNativeDeeplink(final AppAction action) async {
+    final uri = await action.getNativeUri();
+    return _launcherUtil.canLaunchUrl(uri);
   }
 }