+++ lib/src/core/enums/enums.dart
@@ -1,2 +1,2 @@
 export 'action_type_enum.dart';
-export 'platform_enum.dart';
+export 'platform_type.dart';