+++ test/deeplink_x_test.dart
@@ -17,7 +17,7 @@ class DummyAppAction extends AppAction {
 }
 
 class DummyStoreAppAction extends StoreAppAction {
-  DummyStoreAppAction() : super(actionType: DummyActionType.open, platform: PlatformEnum.android);
+  DummyStoreAppAction() : super(actionType: DummyActionType.open, platform: PlatformType.android);
 
   @override
   Future<Uri> getNativeUri() async => Uri(scheme: 'test', host: 'test');
@@ -66,6 +66,16 @@ void main() {
         expect(action, isA<AppAction>());
       });
 
+      test('WhatsApp', () {
+        final action = WhatsApp.open();
+        expect(action, isA<AppAction>());
+      });
+
+      test('LinkedIn', () {
+        final action = WhatsApp.open();
+        expect(action, isA<AppAction>());
+      });
+
       test('IOSAppStore', () {
         const action = IOSAppStore.open;
         expect(action, isA<AppAction>());
@@ -86,6 +96,11 @@ void main() {
         expect(action, isA<AppAction>());
       });
 
+      test('MyketStore', () {
+        const action = MyketStore.open;
+        expect(action, isA<AppAction>());
+      });
+
       test('MacAppStore', () {
         const action = MacAppStore.open;
         expect(action, isA<AppAction>());
@@ -98,9 +113,9 @@ void main() {
     });
 
     group('Enums:', () {
-      test('PlatformEnum', () {
-        const platform = PlatformEnum.android;
-        expect(platform, isA<PlatformEnum>());
+      test('PlatformType', () {
+        const platform = PlatformType.android;
+        expect(platform, isA<PlatformType>());
       });
     });