+++ test/src/utils/platform_util_test.dart
@@ -1,0 +1,2 @@
+// TODO: Add tests for platform_util.dart
+void main() {}