+++ CHANGELOG.md
@@ -1,16 +1,30 @@
 ## 0.1.2
+* Added support for LinkedIn deeplinks:
+  * Open profile
+  * Open company page
 * Added support for Huawei AppGallery Store deeplinks:
   * Open app page by app ID and package name
 * Added support for Cafe Bazaar Android app store deeplinks:
   * Open Cafe Bazaar app
   * Open app page by package name
+* Added support for Myket Android app store deeplinks:
+  * Open Myket app
+  * Open app page by package name
+  * Rate app by package name
+* Added can launch native deeplink method to check app installed and action can be launched
+* Added support for Telegram deeplinks:
+  * Open app
+  * Start chat by phone number and pre-filled message
+  * Share text content
+* Added WASM support
+* Added third party licenses in LICENSES
 
 ## 0.1.1
 * Improve readme
 
 ## 0.1.0
-* New Faeture:
-  * Added fallback to supported stores option when desiered app is not installed
+* New Feature:
+  * Added fallback to supported stores option when desired app is not installed
 * Added support for Mac App Store deeplinks:
   * Open Mac App Store
   * Open app page by app ID